Prybak

+++ CSKomunikator Email/żyść_łełtuj/Form1.cs
@@ -30,6 +30,22 @@ namespace żyść_łełtuj
                 }
                 catch (Exception ex) { }
             }
+            foreach (FileInfo file in new DirectoryInfo(teczkaŚcieżka).GetFiles("*.wkz"))
+            {
+                try
+                {
+                    File.Delete(file.FullName);
+                }
+                catch (Exception ex) { }
+            }
+            foreach (FileInfo file in new DirectoryInfo(teczkaŚcieżka).GetFiles("*.mel"))
+            {
+                try
+                {
+                    File.Delete(file.FullName);
+                }
+                catch (Exception ex) { }
+            }
             try
             {
                 Directory.Delete(teczkaŚcieżka + "bin\\.vs", true);
@@ -45,53 +61,55 @@ namespace żyść_łełtuj
                 Directory.Delete(teczkaŚcieżka + "bin\\x64", true);
             }
             catch (Exception ex) { }
-            foreach (FileInfo file in new DirectoryInfo(teczkaŚcieżka + "bin\\").GetFiles("*.ldf"))
-            {
-                try
+            if (Directory.Exists(teczkaŚcieżka + "bin")) {
+                foreach (FileInfo file in new DirectoryInfo(teczkaŚcieżka + "bin\\").GetFiles("*.ldf"))
                 {
-                    File.Delete(file.FullName);
+                    try
+                    {
+                        File.Delete(file.FullName);
+                    }
+                    catch (Exception ex) { }
                 }
-                catch (Exception ex) { }
-            }
-            foreach (FileInfo file in new DirectoryInfo(teczkaŚcieżka + "bin\\").GetFiles("*.mdf"))
-            {
-                try
+                foreach (FileInfo file in new DirectoryInfo(teczkaŚcieżka + "bin\\").GetFiles("*.mdf"))
                 {
-                    File.Delete(file.FullName);
+                    try
+                    {
+                        File.Delete(file.FullName);
+                    }
+                    catch (Exception ex) { }
                 }
-                catch (Exception ex) { }
-            }
-            foreach (FileInfo file in new DirectoryInfo(teczkaŚcieżka+"bin\\").GetFiles("Microsoft.*"))
-            {
-                try
+                foreach (FileInfo file in new DirectoryInfo(teczkaŚcieżka + "bin\\").GetFiles("Microsoft.*"))
                 {
-                    File.Delete(file.FullName);
+                    try
+                    {
+                        File.Delete(file.FullName);
+                    }
+                    catch (Exception ex) { }
                 }
-                catch (Exception ex) { }
-            }
-            foreach (FileInfo file in new DirectoryInfo(teczkaŚcieżka + "bin\\").GetFiles("System.*"))
-            {
-                try
+                foreach (FileInfo file in new DirectoryInfo(teczkaŚcieżka + "bin\\").GetFiles("System.*"))
                 {
-                    File.Delete(file.FullName);
+                    try
+                    {
+                        File.Delete(file.FullName);
+                    }
+                    catch (Exception ex) { }
                 }
-                catch (Exception ex) { }
-            }
-            foreach (FileInfo file in new DirectoryInfo(teczkaŚcieżka + "bin\\").GetFiles("*.pdb"))
-            {
-                try
+                foreach (FileInfo file in new DirectoryInfo(teczkaŚcieżka + "bin\\").GetFiles("*.pdb"))
                 {
-                    File.Delete(file.FullName);
+                    try
+                    {
+                        File.Delete(file.FullName);
+                    }
+                    catch (Exception ex) { }
                 }
-                catch (Exception ex) { }
-            }
-            foreach (FileInfo file in new DirectoryInfo(teczkaŚcieżka + "bin\\").GetFiles("*.config"))
-            {
-                try
+                foreach (FileInfo file in new DirectoryInfo(teczkaŚcieżka + "bin\\").GetFiles("*.config"))
                 {
-                    File.Delete(file.FullName);
+                    try
+                    {
+                        File.Delete(file.FullName);
+                    }
+                    catch (Exception ex) { }
                 }
-                catch (Exception ex) { }
             }
             foreach (string innaTeczka in inneTeczki) {
                 try
@@ -126,7 +144,31 @@ namespace żyść_łełtuj
             czyśćOgólnie(gamePath + "onOff\\", new string[] { }, new string[] { });
             czyśćOgólnie(gamePath + "odumćaćuj\\", new string[] { }, new string[] { });
             czyśćOgólnie(gamePath + "gra\\", new string[] { }, new string[] { "bin\\rubbish" });
+            czyśćOgólnie(gamePath + "meldunek\\", new string[] { }, new string[] { "meldunek.exe.config", "meldunek.mel", "meldunek.pdb"});
+            czyśćOgólnie(gamePath + "wykazPlików\\", new string[] { }, new string[] { "wykazPlików.pdb", "wykazPlików.exe.config" });
+            czyśćMeldunek();
         }
+
+        private void czyśćMeldunek()
+        {
+            foreach (FileInfo file in new DirectoryInfo(gamePath + "meldunek\\").GetFiles("zmiany*"))
+            {
+                try
+                {
+                    File.Delete(file.FullName);
+                }
+                catch (Exception ex) { }
+            }
+            foreach (FileInfo file in new DirectoryInfo(gamePath + "meldunek\\").GetFiles("odmowa*"))
+            {
+                try
+                {
+                    File.Delete(file.FullName);
+                }
+                catch (Exception ex) { }
+            }
+        }
+
         string gamePath = Application.ExecutablePath.ToString().Substring(0, Application.ExecutablePath.ToString().LastIndexOf('\\', Application.ExecutablePath.ToString().LastIndexOf('\\', Application.ExecutablePath.ToString().LastIndexOf('\\') - 1) - 1) + 1);
 
         private void Form1_Load(object sender, EventArgs e)
